Yes.
Yes.

It works.

It loads, it saves, it does all that stuff.

It works, even if federico complains that we did not test close.

Repetition, alarms, all that stuff you all guys love.

It it is there.  We did minimal testing, but we know you will
happilly commit a fix if you find a problem, right?

Ok, we are off to a party now.

Miguel