This app is going to break on weird versions of "w". If you have a weird
This app is going to break on weird versions of "w". If you have a
weird version of w (i.e. not Linux or Solaris), try it and if it
breaks send me some sample "w" output. (well, it might break on Linux
and Solaris too...)

Sun Apr 26 02:31:39 1998  Havoc Pennington  <hp@pobox.com>

	* gw.c: OK, the configurable popup menus now work for me. You can
	define a command with %u (user) and %t (tty) escape sequences,
	then right-click on the user in the list to run the command with
	their username and tty substituted. e.g. "xterm -e write %u %t".