I'm putting grun back to its pre-Gnomified state. Why? The point of grun
I'm putting grun back to its pre-Gnomified state. Why? The point of grun
is to provide the equivalent of a command line without having to run an
xterm. We don't need a GnomeDialog. We don't need an "OK" and "Cancel"
button. It's not meant to be user-friendly - people who want
user-friendliness don't use command lines, remember? :) It's meant to be
convenient, which means it should take as little screen space, pop up as
fast as possible, and get out of the way when you're done. Adding buttons
and a gnome dialog contradicts this goal.