I've added the gnomecard utility to the gncal sub-directory. This is a
I've added the gnomecard utility to the gncal sub-directory. This is a work
in progress.

The only missing part is the main view of the utility so, if anybody wants
to contribute with suggestions, they are very wellcomed. I plan to display
the vCard business cards in two modes: a "card mode", using Federico's
canvas widget (this will enable the util to print the cards) and a "tree
mode" using gtk_tree (this one will be very cool and practical, I think).

Arturo