Added a documentation README file Updated the changelog, shrunk the TODO list
Splash now runs correctly in a non X11 environment (it just runs the program
script in such a way as splac will behave). Its file descriptor handling
bug is fixed. In an X11 environment it captures stderr and if there is
data at the end of the run the user gets a nice dialog box with some diagnostics
in it.

Still 100% gtk - after all someone may want to use this to install Gnome ;)