Ok this is good old "dialog" with Gnome support.
Gnome mode is autoenabled when $DISPLAY is set. All the widgets are implemented
gauge. Most of them work too

Checklist.c needs fixing - it wants to walk the selected list of widgets
for a CList - really it looks like there should be a

	GList *gtk_clist_selected_rows(clist);

to return the _number_ of the rows selected

Also it wont build without ncurses, it doesnt build without WITH_GNOME set
in dialog.h and the autoconf stuff doesnt look up.

If someone could look at the gtk_clist item and the autoconf deep magic I'll
fix the general mess over the next few days

Alan