Started a rewrite of the GUI for gnomecard.
The first step is to replace the tree view with a list view,
and to replace the business card on the right side with a list of
the fields from the currently selected record.

I will move to the card edit screen(s) next.

At the moment it builds, but it will not work as well as before. I hope
to have it at least functionally equivalent to the original in the
next few days.

Dr Mike <msf@redhat.com>