Don't redefine _pam_overwrite if already defined.
This is a security critical file. I'm using multiple commits which small
changes between them so it's easier to run a diff and find possible
security holes.

The original code is in the `sh_utils' branch and tagged `sh_utils_1_16'.

Currently this is a /bin/su that uses glib instead of its own xmalloc() etc.

Martin