Improve GDict debug messages handling
This commit changes the debug messages handling in GDict, using a system
similar to the one GLib and GTK+ use. You can either set the GDICT_DEBUG
environment variable to a colon-separated or comma-separated list of
debug flags, or you can use the --gdict-debug and --gdict-no-debug
command line arguments to switch on (and off) the debug flags. The
GDICT_NOTE() macro has been added as well, replacing _gdict_debug() for
a more comprehensive debug output.

The entire debug message system is enabled either by calling the
gdict_debug_init() function or by using the GOptionGroup that is
returned by gdict_get_option_group().

The debug message system is also depending on whether the GDict library
has been compiled with the GDICT_ENABLE_DEBUG flag -- which is the
default for development/unstable cycles.

svn path=/trunk/; revision=7661