[build] Remove gfloppy
The migration to Git broke Gfloppy, which was included as a
svn:external.

In order to keep gnome-utils running we need to remove the
various checks we do for Gfloppy from the configure and
avoid recursing into /gfloppy.

This commit essentially removes Gfloppy from the gnome-utils
package until such a time when we have decided whether we
drop it completely or we keep it as a sub-module directly
inside the repository.