[build] Massive update to the configure template
Let's bring the configure.ac template into the new century, shall we?

Use macros like AS_IF and AS_CASE instead of shell-like if and case.

Cut down the checks on some crap that we don't use anymore or that
comes with being a GNOME project.

Clear up the coding style.

Add options for the anal-retentive compiler flags and be more
aggressive with the debug defines.

Clean the summary.