configure: Check for shape.h if the xext is available instead of if not
Previously the configure script would only check for the shape.h
header if the xext pkg-config package is *not* available. This meant
that gnome-screenshot would only use the shape extension if the xext
package isn't installed which doesn't make much sense. Presumably the
intention was to check for the package and then check for the header
so that the shape extension is only used if both are available. This
patch fixes it to do that.

https://bugzilla.gnome.org/show_bug.cgi?id=620100